requirements.txt
# ==================== ML / DATA SCIENCE ====================
catboost==1.2.7
scikit-learn==1.5.2
scipy==1.14.1
pandas==2.2.3
numpy==1.26.4

# ==================== PERFORMANCE ====================
numba==0.60.0

# ==================== VISUALIZATION ====================
matplotlib==3.9.2
seaborn==0.13.2

# ==================== CONFIG / UTILS ====================
pyyaml==6.0.2
tqdm==4.66.5

# ==================== TESTING ====================
pytest==8.3.3
pytest-cov==5.0.0
pytest-timeout==2.3.1

# ==================== OPTIONAL ====================
# onnxruntime==1.19.2